+++ src/root/RouterList.js
@@ -31,7 +31,6 @@ const Products = lazy(() => import('../components/Products/Products'));
 const ProductIsCheck = lazy(() => import('../components/Products/ProductIsCheck/ProductIsCheck'));
 const ProductEditPage = lazy(() => import('../components/Products/AddingProductEdit/ProductsEditPage'));
 const ProductLocationsList = lazy(() => import('../components/Products/ProductLocationsList/ProductLocationsList'));
-const ProductLocationsCopy = lazy(() => import('../components/Products/productListCopy/ProductLocationsCopy'));
 const ProductsPageOne = lazy(() => import('../components/Products/AddingProductPageOne/ProductsPageOne'));
 
 // -----------------------------Authentication-----------------
@@ -175,16 +174,12 @@ export default function RouterList() {
               <ProductEditPage />
             </Suspense>
           } />
+
           <Route path="/products/location" element={
-            <Suspense fallback={<div className="w-full h-full"><LoadingForSeller /></div>}>
-              <ProductLocationsCopy />
-            </Suspense>
-          } />
-          {/* <Route path="/products/location" element={
             <Suspense fallback={<div className="w-full h-full"><LoadingForSeller /></div>}>
               <ProductLocationsList />
             </Suspense>
-          } /> */}
+          } />
           <Route path="/products/location/add/:id" element={
             <Suspense fallback={<div className="w-full h-full"><LoadingForSeller /></div>}>
               <ProductsPageOne />
